+++ xcloud-components-core/src/main/java/com/wl4g/components/core/bean/erm/DockerCluster.java
@@ -23,17 +23,15 @@ public class DockerCluster extends BaseBean {
 
 	private static final long serialVersionUID = -7546448616357790576L;
 
-	private List<Integer> hostIds;
-
+	private List<Long> hostIds;
 	private String name;
-
 	private String masterAddr;
 
-	public List<Integer> getHostIds() {
+	public List<Long> getHostIds() {
 		return hostIds;
 	}
 
-	public void setHostIds(List<Integer> hostIds) {
+	public void setHostIds(List<Long> hostIds) {
 		this.hostIds = hostIds;
 	}
