+++ xcloud-components-core/src/main/java/com/wl4g/components/core/bean/iam/Department.java
@@ -20,7 +20,7 @@ import com.wl4g.components.core.bean.BaseBean;
 public class Department extends BaseBean {
 	private static final long serialVersionUID = -2227319200968900422L;
 
-	private Integer groupId;
+	private Long groupId;
 
 	private String displayName;
 
@@ -28,11 +28,11 @@ public class Department extends BaseBean {
 
 	private String contactPhone;
 
-	public Integer getGroupId() {
+	public Long getGroupId() {
 		return groupId;
 	}
 
-	public void setGroupId(Integer groupId) {
+	public void setGroupId(Long groupId) {
 		this.groupId = groupId;
 	}
